+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Change Log
 
 ### Fixes
 
+- Fix an issue where a vertical orientation document broke the layout on printing.
 - Fix an issue where the syntax highlighting indicator could display twice.
 - Fix an issue where the separator was selected meaninglessly in the Window pane if the window tabbing setting was set to “Manually”.
 - Fix an issue where editor's text orientation was not cascaded to the print operation when the window was restored from the last session.